Minicluster: Started to work on isolating ports and container names
Summary:
We need to randomize port allocation, and remove a few assumptions in the
container creation logic. In order to do this, we also need to move the multi-instance
port allocation logic for ease of access.
This is the first step in being able to run multiple miniclusters at the
same time.

The `isolate` function isn't being called quite yet; this diff only declares new config
variables that will be overwritten in future diffs. This commit will have no behavior change, and
is only setting up the field for future diffs that will have isolation.

Test Plan:
Ran the following successfully:
```
$ peloton::setup
$ peloton::respool::create
$ peloton::job::stateless::create
```

Reviewers: adityacb, apoorvaj, amitbose, O521 Project Peloton: Add reviewers, rihan

Reviewed By: adityacb, O521 Project Peloton: Add reviewers, rihan

Subscribers: rihan, jenkins

Differential Revision: https://code.uberinternal.com/D3309559