Update desired host-pool with current host-pool
Summary:
When the current host-pool is changed directly (for
example an API call), also update the desired host-pool.
Otherwise the host goal-state may try to revert the
current pool.

Test Plan:
Manually verified that desired host-pool is set to the
same value as current host-pool on ChangePool API and
when a new host is added.

Reviewers: O521 Project Peloton: Add reviewers, sachins

Reviewed By: O521 Project Peloton: Add reviewers, sachins

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D3318981