Minicluster: Port isolation finished
Summary:
In this commit, we move some of the port calculations to the `__init__`
function of the minicluster. When we decide to actually `setup` the
cluster, we will then use those pre-determined values.
There is still work to be done with respect to the naming of containers.
We need to make sure the container names from the miniclusters don't
collide.

NOTE: The `isolate` function isn't called by default here; so no behavior
change on master.

Reviewers: adityacb, amitbose, apoorvaj, O521 Project Peloton: Add reviewers

Reviewed By: adityacb, O521 Project Peloton: Add reviewers

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D3318207