Fix hostinfo entry creation
Summary:
Before, an entry was created in host_info table in `StartMaintenance`. This was correct when only hosts in maintenance were persisted in cassandra. Now, with all hosts being persisted in cassandra, we should create the host entry when the host is discovered.

This diff addresses the above issue by creating an entry in host_info table when AgentMap is populated and only updating the host goal state in `StartMaintenance`.

Test Plan: Unit tests

Reviewers: amitbose, varung, O521 Project Peloton: Add reviewers

Reviewed By: varung, O521 Project Peloton: Add reviewers

Subscribers: jenkins

Maniphest Tasks: T3968077

Differential Revision: https://code.uberinternal.com/D3323843