Use timestamp datatypes in events table
Summary:
Reading timestamp_uuid is very expensive, so this diff adds timestamp column which is more performant in read path.

Reason to use UUID was to prevent conflicts if two timestamps were same, probability of this happening is low and event if it happens then there is no downside.

This diff adds in write path, which will be rolled out as patch to stateless clusters and after two weeks we will revert to use timestamp on read path.

{F162794407}

Reviewers: O521 Project Peloton: Add reviewers, adityacb, sishi

Reviewed By: O521 Project Peloton: Add reviewers, adityacb

Subscribers: apoorvaj, jenkins

Differential Revision: https://code.uberinternal.com/D3330693