Remove persisting of pod_status with pod_events
Summary: Pod Status is persisted with each pod event, whereas not used in the read path. This diff removes the part to persist pod_status. I do not plan to remove pod_status field from pod_events and we can revert back to adding once we see the necessity. As pod events is purged regularly, not having pod_status history will not be a problem.

Reviewers: O521 Project Peloton: Add reviewers, adityacb, sishi

Reviewed By: O521 Project Peloton: Add reviewers, adityacb

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D3330723