Reconcile host-pools from host_info table
Summary:
AgentMap was being used to determine the list
of hosts in the system. However, this list only
contains currently registered hosts; hosts in
maintenance or temporarily disconnected are
not present. The host-pool for such hosts was
getting cleared out because of periodic
reconciliation in host-pool manager.
Now we get the list of hosts from DB which
contains hosts in all different states.

Test Plan:
Updated unit-tests. Put host into maintenance
and verified that its host-pool information
in the DB was not wiped out. Removed a host
from DB and verified that cached pool information
gets updated.

Reviewers: O521 Project Peloton: Add reviewers, sachins, varung

Reviewed By: O521 Project Peloton: Add reviewers, varung

Subscribers: jenkins

Maniphest Tasks: T4002377

Differential Revision: https://code.uberinternal.com/D3337677