Refresh default pool capacity on delete host-pool
Summary:
Hosts from the deleted pool are added to the default
pool, so we need to update the capacity of that
pool.

Also removed the recover() method because reconcile()
already does the same thing (and more).

Test Plan:
Unit-tests. Ran integration test for host-pool capacity
many time; prior to this change it failed about 50% of
times, now it works always.

Reviewers: O521 Project Peloton: Add reviewers, varung, sachins

Reviewed By: O521 Project Peloton: Add reviewers, sachins

Maniphest Tasks: T4053917

Differential Revision: https://code.uberinternal.com/D3364015